requirements.txt
numpy==2.4.0
openmeteo_requests==1.7.4
pandas==2.3.3
requests-cache==1.2.1
retry-requests==2.0.0
psycopg2-binary==2.9.11